Reduce nodeLimit for PerfTest, introduce FullPerftTests
